+++ app/src/main/java/com/sofato/krone/ui/expenses/EditExpenseScreen.kt
@@ -317,8 +317,8 @@ fun EditExpenseScreen(
 
     if (showDatePicker) {
         val today = LocalDate.today()
-        val todayMillis = today.toEpochDays().toLong() * 86_400_000L
-        val selectedMillis = (selectedDate ?: today).toEpochDays().toLong() * 86_400_000L
+        val todayMillis = today.toEpochDays() * 86_400_000L
+        val selectedMillis = (selectedDate ?: today).toEpochDays() * 86_400_000L
         val datePickerState = rememberDatePickerState(
             initialSelectedDateMillis = selectedMillis,
             selectableDates = object : SelectableDates {
